+++ sample/src/com/viewpagerindicator/sample/SampleTabsDefault.java
@@ -6,7 +6,6 @@ import android.support.v4.app.FragmentManager;
 import android.support.v4.view.ViewPager;
 import com.viewpagerindicator.sample.R;
 import com.viewpagerindicator.TabPageIndicator;
-import com.viewpagerindicator.TitleProvider;
 
 public class SampleTabsDefault extends BaseSampleActivity {
 	private static final String[] CONTENT = new String[] { "Recent", "Artists", "Albums", "Songs", "Playlists", "Genres" };
@@ -25,7 +24,7 @@ public class SampleTabsDefault extends BaseSampleActivity {
 		mIndicator.setViewPager(mPager);
 	}
 	
-	class GoogleMusicAdapter extends TestFragmentAdapter implements TitleProvider {
+	class GoogleMusicAdapter extends TestFragmentAdapter {
 		public GoogleMusicAdapter(FragmentManager fm) {
 			super(fm);
 		}
@@ -41,7 +40,7 @@ public class SampleTabsDefault extends BaseSampleActivity {
 		}
 
 		@Override
-		public String getTitle(int position) {
+		public CharSequence getPageTitle(int position) {
 			return SampleTabsDefault.CONTENT[position % SampleTabsDefault.CONTENT.length].toUpperCase();
 		}
 	}